+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bugfixes
 
  - Fix dependency check ([#422](https://github.com/kizniche/mycodo/issues/422))
+ - Try lower integration times when TSL2561 sensor is saturated ([#450](https://github.com/kizniche/mycodo/issues/450))
 
 ## 6.0.4 (2018-04-21)